MDW-1645: using doubles for zoom values. Fixing property name.

+++ src/main/java/com/tomtom/gwt/mapbox/gl/client/layers/style/ZoomAndValue.java
@@ -7,13 +7,16 @@ import jsinterop.annotations.JsProperty;
 import jsinterop.annotations.JsType;
 
 /**
- *
+ * Zoom-and-property functions allow the appearance of a map feature to change with both its properties and zoom. 
+ * Each stop is an array with two elements, the first is an object with a property input
+ * value and a zoom, and the second is a function output value.
+ * @see https://www.mapbox.com/mapbox-gl-js/style-spec/#types-function Zoom-and-property functions
  */
 @JsType(isNative = true, name = JS_OBJECT_TYPE, namespace = JsPackage.GLOBAL)
 public class ZoomAndValue {
-    
+
     @JsOverlay
-    public static final ZoomAndValue build(Integer zoom, Integer value) {
+    public static final ZoomAndValue build(Double zoom, Double value) {
         ZoomAndValue result = new ZoomAndValue();
         if (zoom != null) {
             result.setZoom(zoom);
@@ -23,13 +26,13 @@ public class ZoomAndValue {
         }
         return result;
     }
-    
+
     private ZoomAndValue() {
     }
-    
+
     @JsProperty
-    private native void setZoom(int zoom);
-    
+    private native void setZoom(double zoom);
+
     @JsProperty
-    private native void setValue(int value);
+    private native void setValue(double value);
 }
