+++ src/main/java/com/tomtom/gwt/mapbox/gl/client/mapoptions/MapOptions.java
@@ -76,13 +76,13 @@ public final class MapOptions {
          * @param value (default 0) The minimum zoom level of the map (1-20).
          */
         @JsProperty
-        public native void setMinZoom(int value);
+        public native void setMinZoom(double value);
 
         /**
          * @param value (default 20) The maximum zoom level of the map (1-20).
          */
         @JsProperty
-        public native void setMaxZoom(int value);
+        public native void setMaxZoom(double value);
 
         /**
          * @param style The map's Mapbox style. This must be an a JSON object conforming to the schema described in the Mapbox Style Specification , or a URL to such JSON. To load a style from the
